requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.0
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
